Used classes as local namespaces for hooks with the same name in the same module.
