--- src/pytask/_trace.py
@@ -1,19 +1,0 @@
-import functools
-import pdb
-
-import pytask
-
-
-@pytask.hookimpl(hookwrapper=True)
-def pytask_execute_task(task):
-    if hasattr(task, "function"):
-        task.function = wrap_function_for_tracing(task.function)
-    yield
-
-
-def wrap_function_for_tracing(function):
-    @functools.wraps(function)
-    def wrapper(*args, **kwargs):
-        pdb.runcall(function, *args, **kwargs)
-
-    return wrapper
